Split and refactor remote data subscriptions
