+++ src/components/Dialog/ManageSigners.tsx
@@ -9,8 +9,8 @@ import PersonAddIcon from "@material-ui/icons/PersonAdd"
 import { Account } from "../../context/accounts"
 import { trackError } from "../../context/notifications"
 import { useAccountData } from "../../hooks"
-import { ObservedAccountData } from "../../lib/subscriptions"
 import { createTransaction } from "../../lib/transaction"
+import { ObservedAccountData } from "../../subscriptions"
 import { Box, HorizontalLayout } from "../Layout/Box"
 import ManageSignersForm, { SignerUpdate } from "../ManageSigners/ManageSignersForm"
 import TransactionSender from "../TransactionSender"
