+++ src/components/ManageSigners/ManageSignersForm.tsx
@@ -7,9 +7,9 @@ import Tooltip from "@material-ui/core/Tooltip"
 import CheckIcon from "@material-ui/icons/Check"
 import CloseIcon from "@material-ui/icons/Close"
 import InfoIcon from "@material-ui/icons/Info"
-import { renderFormFieldError } from "../../lib/errors"
-import { ObservedAccountData } from "../../lib/subscriptions"
 import { trackError } from "../../context/notifications"
+import { renderFormFieldError } from "../../lib/errors"
+import { ObservedAccountData } from "../../subscriptions"
 import { ActionButton, DialogActionsBox } from "../Dialog/Generic"
 import { Box, HorizontalLayout, VerticalLayout } from "../Layout/Box"
 import SignersEditor from "./SignersEditor"
