+++ src/components/TransactionSummary/TransactionSummary.tsx
@@ -4,7 +4,7 @@ import { Memo, Transaction } from "stellar-sdk"
 import Divider from "@material-ui/core/Divider"
 import { useAccountData } from "../../hooks"
 import { Account, AccountsContext } from "../../context/accounts"
-import { ObservedAccountData } from "../../lib/subscriptions"
+import { ObservedAccountData } from "../../subscriptions"
 import { List, ListItem } from "../List"
 import PublicKey from "../PublicKey"
 import OperationListItem from "./Operations"
